refactor: log export
