--- src/views/dashboard/logs/query/ExportLog.vue
@@ -1,73 +1,0 @@
-<template lang="pug">
-a-button(
-  type="outline"
-  size="small"
-  :disabled="!queryState.tableName"
-  @click="exportSql"
-)
-  template(#icon)
-    svg.icon
-      use(href="#export")
-  | {{ $t('dashboard.exportCSV') }}
-</template>
-
-<script setup name="ExportLog" lang="ts">
-  import { storeToRefs } from 'pinia'
-  import fileDownload from 'js-file-download'
-  import editorAPI from '@/api/editor'
-  import useLogsQueryStore from '@/store/modules/logs-query'
-  import { getWhereClause, parseOrderBy } from './until'
-
-  interface TSColumn {
-    name: string
-    multiple?: number
-  }
-
-  interface Column {
-    name: string
-    data_type: string
-  }
-
-  const props = defineProps<{
-    columns: Column[]
-    tsColumn: TSColumn | null
-  }>()
-
-  const { queryState } = useLogsQueryStore()
-  const { finalQuery } = storeToRefs(useLogsQueryStore())
-
-  function getExportSql() {
-    if (props.columns && props.columns.length > 0) {
-      try {
-        let fields = []
-        if (props.tsColumn && props.tsColumn.name) {
-          fields.push(props.tsColumn.name)
-        }
-        fields = [
-          ...fields,
-          ...props.columns.map((column) => column?.name).filter((v) => v && v !== props.tsColumn?.name),
-        ]
-        const where = getWhereClause(finalQuery.value)
-        const order = parseOrderBy(finalQuery.value) || 'desc'
-        const limit = 10000
-        return `SELECT ${fields.join(', ')} FROM ${queryState.tableName} ${where} ORDER BY ${
-          props.tsColumn?.name || 'id'
-        } ${order} LIMIT ${limit}`
-      } catch (error) {
-        return `SELECT * FROM ${queryState.tableName} LIMIT 10000`
-      }
-    }
-    return `SELECT * FROM ${queryState.tableName} LIMIT 10000`
-  }
-
-  function exportSql() {
-    if (!queryState.tableName) {
-      return
-    }
-    editorAPI.runSQLWithCSV(getExportSql()).then((result) => {
-      fileDownload(result as unknown as string, `${queryState.tableName}.csv`)
-    })
-  }
-</script>
-
-<style scoped lang="less"></style>
